requirements/base.txt
django==3.2.4

django-configurations==2.2
dj-database-url==0.5.0
